❤️ TEST-#82: Split plugin tests into tests/plugin/ mirroring the source package
